Move SubscribeOptions into its own file
