+++ event-store/src/adapters/mod.rs
@@ -5,18 +5,3 @@ mod store;
 pub use self::cache::{CacheResult, PgCacheAdapter};
 pub use self::emitter::AmqpEmitterAdapter;
 pub use self::store::{PgQuery, PgStoreAdapter, SaveResult, SaveStatus};
-
-#[derive(Debug, Clone)]
-pub struct SubscribeOptions {
-    pub replay_previous_events: bool,
-    pub save_on_receive: bool,
-}
-
-impl Default for SubscribeOptions {
-    fn default() -> Self {
-        Self {
-            replay_previous_events: true,
-            save_on_receive: true,
-        }
-    }
-}
