+++ event-store/src/prelude.rs
@@ -1,7 +1,7 @@
-pub use crate::adapters::SubscribeOptions;
 pub use crate::aggregator::Aggregator;
 pub use crate::event::Event;
 pub use crate::event_context::EventContext;
 pub use crate::event_handler::EventHandler;
 pub use crate::store::Store;
 pub use crate::store_query::StoreQuery;
+pub use crate::subscribe_options::SubscribeOptions;
